requirements.txt
pytest==8.1.1
pandas==2.2.1
boto3==1.34.79
psycopg2-binary==2.9.9
